Rename `file-index` to `plugin-index`
