+++ lib/plugins/plugin-index-thread.js
