+++ lib/plugins/plugin-index.js
