+++ lib/plugins/test/plugin-index-test.js
@@ -1,6 +1,6 @@
 // # file-index-test.js
 import { expect } from 'chai';
-import Index from '../file-index.js';
+import Index from '../plugin-index.js';
 import { FileType } from 'sc4/core';
 import { resource } from '#test/files.js';
