+++ lib/plugins/worker-pool.js
@@ -8,10 +8,10 @@ import BaseWorkerPool from 'sc4/threading/worker-pool.js';
 export default class WorkerPool extends BaseWorkerPool {
 	constructor(opts) {
 		if (sea.isSea()) {
-			const script = sea.getAsset('threads/file-index-thread.js', 'utf8');
+			const script = sea.getAsset('threads/plugin-index-thread.js', 'utf8');
 			super({ ...opts, script });
 		} else {
-			const url = import.meta.resolve('./file-index-thread.js');
+			const url = import.meta.resolve('./plugin-index-thread.js');
 			super({ ...opts, url });
 		}
 	}
